Extract desktop/src/settings-store.js (#17 Stage 4a)
Carve the pure file-IO and settings-shape helpers out of
desktop/src/main.js into a focused module:

  ensureDir, readJson, writeJson, readText,
  backupFileOncePerMinute, nonEmptyChatSessions, sanitizedSettings

main.js destructure-imports them from settings-store, so all existing
call sites (including settings(), persistAuthSecrets, runValidation,
buildProjectIndex, etc.) keep working unchanged. The smoke regex for
`delete next.auth.openaiApiKey` now reads settings-store.js since the
sanitizer moved.

Stage 4 of issue #17's plan called for a four-module split of main.js
(projects + settings-store + runs + ipc). This PR ships the cleanest
seam — the pure helpers — as a stand-alone increment. The remaining
project-management, run-lifecycle, and IPC-wiring extracts are
follow-ups so each lands as its own reviewable change.

Numbers:
- main.js: 1888 -> 1843 (-45)
- settings-store.js: 69 lines (new)

Verified:
- npm run check: clean
- npm run smoke: 71 unit tests + 12 behavior smoke tests pass

Refs #17.
